Add base class for next and previous view model
