+++ src/AudioBand/ViewModels/PreviousButtonVM.cs
@@ -1,136 +1,33 @@
-﻿using System;
-using System.Diagnostics;
-using System.Drawing;
-using AudioBand.Models;
+﻿using AudioBand.Models;
 using AudioBand.Resources;
 using AudioBand.Settings;
 
-#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 namespace AudioBand.ViewModels
 {
     /// <summary>
     /// View model for the previous button.
     /// </summary>
-    public class PreviousButtonVM : ViewModelBase<PreviousButton>
+    public class PreviousButtonVM : PlaybackButtonVMBase<PreviousButton>
     {
-        private readonly IAppSettings _appsettings;
-        private readonly IResourceLoader _resourceLoader;
-        private IImage _image;
-
         /// <summary>
         /// Initializes a new instance of the <see cref="PreviousButtonVM"/> class.
         /// </summary>
-        /// <param name="appsettings">The app settings.</param>
+        /// <param name="appSettings">The app settings.</param>
         /// <param name="resourceLoader">The resource loader.</param>
         /// <param name="dialogService">The dialog service</param>
-        public PreviousButtonVM(IAppSettings appsettings, IResourceLoader resourceLoader, IDialogService dialogService)
-            : base(appsettings.PreviousButton)
+        public PreviousButtonVM(IAppSettings appSettings, IResourceLoader resourceLoader, IDialogService dialogService)
+            : base(appSettings, resourceLoader, dialogService, appSettings.PreviousButton)
         {
-            DialogService = dialogService;
-            _appsettings = appsettings;
-            _resourceLoader = resourceLoader;
-            LoadImage();
-
-            _appsettings.ProfileChanged += AppsettingsOnProfileChanged;
         }
 
-        public IImage Image
+        protected override PreviousButton GetReplacementModel()
         {
-            get => _image;
-            set => SetProperty(ref _image, value, trackChanges: false);
+            return AppSettings.PreviousButton;
         }
 
-        [PropertyChangeBinding(nameof(PreviousButton.ImagePath))]
-        public string ImagePath
+        protected override IImage GetDefaultImage()
         {
-            get => Model.ImagePath;
-            set
-            {
-                if (SetProperty(nameof(Model.ImagePath), value))
-                {
-                    Image = _resourceLoader.TryLoadImageFromPath(value, _resourceLoader.DefaultPreviousImage);
-                }
-            }
-        }
-
-        [PropertyChangeBinding(nameof(PreviousButton.IsVisible))]
-        public bool IsVisible
-        {
-            get => Model.IsVisible;
-            set => SetProperty(nameof(Model.IsVisible), value);
-        }
-
-        [PropertyChangeBinding(nameof(PreviousButton.Width))]
-        [AlsoNotify(nameof(Size))]
-        public int Width
-        {
-            get => Model.Width;
-            set => SetProperty(nameof(Model.Width), value);
-        }
-
-        [PropertyChangeBinding(nameof(PreviousButton.Height))]
-        [AlsoNotify(nameof(Size))]
-        public int Height
-        {
-            get => Model.Height;
-            set => SetProperty(nameof(Model.Height), value);
-        }
-
-        [PropertyChangeBinding(nameof(PreviousButton.XPosition))]
-        [AlsoNotify(nameof(Location))]
-        public int XPosition
-        {
-            get => Model.XPosition;
-            set => SetProperty(nameof(Model.XPosition), value);
-        }
-
-        [PropertyChangeBinding(nameof(PreviousButton.YPosition))]
-        [AlsoNotify(nameof(Location))]
-        public int YPosition
-        {
-            get => Model.YPosition;
-            set => SetProperty(nameof(Model.YPosition), value);
-        }
-
-        /// <summary>
-        /// Gets the location of the button.
-        /// </summary>
-        public Point Location => new Point(Model.XPosition, Model.YPosition);
-
-        /// <summary>
-        /// Gets the size of the button.
-        /// </summary>
-        public Size Size => new Size(Width, Height);
-
-        /// <summary>
-        /// Gets the dialog service
-        /// </summary>
-        public IDialogService DialogService { get; }
-
-        /// <inheritdoc/>
-        protected override void OnReset()
-        {
-            base.OnReset();
-            LoadImage();
-        }
-
-        /// <inheritdoc/>
-        protected override void OnCancelEdit()
-        {
-            base.OnCancelEdit();
-            LoadImage();
-        }
-
-        private void LoadImage()
-        {
-            Image = _resourceLoader.TryLoadImageFromPath(ImagePath, _resourceLoader.DefaultPreviousImage);
-        }
-
-        private void AppsettingsOnProfileChanged(object sender, EventArgs e)
-        {
-            Debug.Assert(IsEditing == false, "Should not be editing");
-            ReplaceModel(_appsettings.PreviousButton);
+            return ResourceLoader.TryLoadImageFromPath(ImagePath, ResourceLoader.DefaultPreviousImage);
         }
     }
 }
-#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
